tests/wpt/tests/svg/animations/svgintegeroptionalinteger-animation-invalid-value-1.html

<!doctype html>
<title>Test SVGIntegerOptionalInteger animation with invalid value: String values in from and to attributes.</title>
<link rel="help" href="https://www.w3.org/TR/2001/REC-smil-animation-20010904/#FromAttribute">
<link rel="help" href="https://www.w3.org/TR/2001/REC-smil-animation-20010904/#ToAttribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<svg id="svg">
  <defs>
    <filter id="effect">
      <feConvolveMatrix id="matrix" order="3 3" kernelMatrix="0 0 0 0 1 0 0 0 0">
        <animate attributeName="order" begin="0s" dur="4s" from="A B" to="C D" />
      </feConvolveMatrix>
    </filter>
  </defs>
  <rect x="10" y="10" width="80" height="80" fill="blue" filter="url(#effect)" />
</svg>

<script>

async_test(t => {
    const svg = document.getElementById("svg");
    const matrix = document.getElementById("matrix")

    window.addEventListener('load', t.step_func(() => {
        svg.setCurrentTime(2);

        requestAnimationFrame(t.step_func_done(() => {
          assert_equals(matrix.orderX.baseVal, 3);
          assert_equals(matrix.orderY.baseVal, 3);
          assert_equals(matrix.orderX.animVal, matrix.orderX.baseVal);
          assert_equals(matrix.orderY.animVal, matrix.orderY.baseVal);
        }));
    }));
  });

</script>
