LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgintegeroptionalinteger-animation-invalid-value-1-expected.txt


FAIL Test SVGIntegerOptionalInteger animation with invalid value: String values in from and to attributes. assert_equals: expected 3 but got 0
